+++ src/main/java/luu/indepth/block/custom/ModDoorBlock.java
@@ -1,4 +1,4 @@
-package luu.indepth.block;
+package luu.indepth.block.custom;
 
 import net.minecraft.block.DoorBlock;
 import net.minecraft.sound.SoundEvent;
